IdC-L03-TP-code/cloud-Infrastructure/data/app-data/requirements.txt
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.5
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.4.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.9.2
numpy==2.0.2
packaging==21.3
pandas==2.2.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.5.0
Werkzeug==2.0.2
psycopg2-binary==2.9.3